Add files via upload
The following describes how different features of the simple virtual machine has been implemented with justifications. To implement the virtual machine, the files in the working directory were first checked to know which of them a valid assembly file is. After which, C# types with IIinstruction as an interface were detected by reflection, and saved in a list. And each time a new simple machine language code needs to be executed, I compare the saved type names and the opcode and create an instance the opcode by reflection. For memory efficiency, I saved the already found instructions in a structure-based list that help me match instance of such instruction with their opcodes for easy access.